% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bib_complete.R
\name{check_bib_complete}
\alias{check_bib_complete}
\title{check_bib_complete}
\usage{
check_bib_complete(substitute_missing, ref_item)
}
\arguments{
\item{substitute_missing}{Boolean to specify, whether missing data such as title, journal, year or authors should be substituted (TRUE, default) or left empty (FALSE).}

\item{ref_item}{A bibentry item containing the data}
}
\value{
A complete bibentry object. Complete means all necessary fields (author, year, journal, title) are available, but may be empty strings.
}
\description{
Checks for missing essential bibliographic data and optionally substitutes with dummy data.
Always throws a warning with reference to the bibentry.
Substitutions are inspired by the APA style guide
(https://apastyle.apa.org/style-grammar-guidelines/references/missing-information)
}
\examples{
\dontrun{
check_bib_complete(substitute_missing = FALSE, ref_item = target_ref)
}
}
