% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_html.R
\name{drop_html}
\alias{drop_html}
\title{drop_html}
\usage{
drop_html(
  title,
  journal,
  authors,
  year,
  cite_key,
  url,
  include_qr,
  output_dir,
  style,
  use_xaringan = FALSE
)
}
\arguments{
\item{title}{The publication's title as string.}

\item{journal}{The Journal's name as string.}

\item{authors}{The authors' names as string. If this is a list, it has to be collapsed to a single string (s. example).}

\item{year}{The publication's publication year as string.}

\item{cite_key}{A string specifying the citation key within the .bib file. If no key is specified, the first entry is used.}

\item{url}{The URL that should be encoded as QR code.}

\item{include_qr}{Character string specifying the way the QR code should be included or if no QR code should be included.
'embed' results in a stand alone <img> tag within the HTML object, other options are ignored for the time being.
'link' (default) creates a PNG of the QR code and stores it in a subfolder of the HTML file's location. The HTML <img> tag links to this file then.
'link_svg' creates a SVG of the QR code and stores it in a subfolder of the HTML file's location. The HTML <img> tag links to this file then.
'none' creates no QR code.}

\item{output_dir}{A string specifying the relative path, where the rendered output files should be stored.}

\item{style}{A string specifying the desired style for the visual citation. Possible values are:
"modern", "classic", "clean", "none". If "none" is given, the returned html can use a custom css file provided by the user.
This custom CSS file must specify styles for <div> classes "top-row", "title-row" and "author-row".}

\item{use_xaringan}{Boolean to specify if an HTML output is intended to be included in an HTML presentation (like e.g. xaringan) or not.
When including the visual citation via htmltools::includeHTML(), the QR code needs to be in a subfolder
relative to the rendered presentation, not relative to the visual citation.}
}
\value{
A htmltools taglist containing the visual citation as HTML representation including style.
}
\description{
Accepts bibliographic information and returns a htmltools tagList for printing/display.
}
\examples{
\dontrun{
vc_html <- drop_html(
  title = bibentry$title,
  cite_key = bibentry$key,
  journal = bibentry$journal,
  year = bibentry$year,
  authors = paste(bibentry$author, collapse = ", ")
)

htmltools::html_print(dropped)
}

}
