% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_name.R
\name{drop_name}
\alias{drop_name}
\title{drop_name}
\usage{
drop_name(
  file = "sample_data/sample.bib",
  cite_key = "collaboration_2019_ApJL",
  export_as = "html",
  max_authors = 3,
  include_qr = TRUE,
  style = "modern"
)
}
\arguments{
\item{file}{A .bib file.}

\item{cite_key}{A string specifying the citation key within the .bib file. If no key is specified, the first entry is used.}

\item{export_as}{A string specifying the desired output format. For now only supports HTML by using "html".}

\item{max_authors}{Integer number of maximum authors to print. If the number of authors exceeds this, the list is cropped accordingly.}

\item{include_qr}{Boolean value, whether to include a QR code (containing the URL to the DOI) next to the visual citation.}

\item{style}{A string specifying the desired style for the visual citation. Possible values are: "modern", "classic", "clean", "none". If "none" is given, the returned html can use a custom css file provided by the user. The custom CSS file must specify the DIV classes "top-row", "title-row" and "author-row".}
}
\value{
A visual citation
}
\description{
Extracts metadata from a .bib file and exports the visual citation in the specified format.
}
\examples{
\dontrun{
dropped <- drop_name(
  file = "sample_data/sample.bib",
  cite_key = NULL,
  export_as = "html",
  max_authors = 3
)

htmltools::html_print(dropped)
}
}
