% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qr.R
\name{generate_qr}
\alias{generate_qr}
\title{generate_qr}
\usage{
generate_qr(url)
}
\arguments{
\item{url}{A string to encode as QR code.}
}
\value{
The encoded QR code as matrix.
}
\description{
Generates a QR code from a supplied string and return as a plot object.
This is a wrapper function around qrcode::qr_code().
}
