% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qr_svg.R
\name{generate_qr_svg}
\alias{generate_qr_svg}
\title{generate_qr_svg}
\usage{
generate_qr_svg(url, cite_key)
}
\arguments{
\item{url}{A string to encode as QR code.}

\item{cite_key}{A string containing the cite_key to allow for correct file naming.}
}
\value{
Nothing. The encoded QR code is stored to disk.
}
\description{
Generates a QR code from a supplied string and temporarily saves it in the tempdir.
}
\examples{
\dontrun{
generate_svg("https://doi.org/10.1007/bf00396886", cite_key = "Eschrich1983")
}

}
