% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_css_styles.R
\name{get_css_styles}
\alias{get_css_styles}
\title{get_css_style}
\usage{
get_css_styles(style)
}
\arguments{
\item{style}{A string specifying the desired style for the visual citation.}
}
\value{
A list of inline css styles for each element of the visual citation.
}
\description{
Accepts bibliographic information and retuns a htmltools taglist for printing/display.
}
\examples{
\dontrun{
css_styles <- get_css_styles(style = "modern")
}

}
