% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_authors.R
\name{manage_authors}
\alias{manage_authors}
\title{manage_authors}
\usage{
manage_authors(authors, max_authors)
}
\arguments{
\item{authors}{Can be a single string or a list of authors.}

\item{max_authors}{Maximum number of authors to be returned from the list.}
}
\value{
A single string with the desired maximum number of authors.
}
\description{
Returns a cleaned and cropped string of authors for the visual citation.
}
\examples{
\dontrun{
article_authors <- c("Anna", "John", "Simon")
manage_authors(article_authors, max_authors = 2)
}

}
