% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_clipboard.R
\name{to_clipboard}
\alias{to_clipboard}
\title{Copies the visual citation into the clipboard}
\usage{
to_clipboard(png_path)
}
\arguments{
\item{png_path}{Absolute path to the png file}
}
\description{
Copies the visual citation as an image into the clipboard. Mac, Linux and Windows are supported. For Windows Powershell is used. For Linux xclip is required.
}
