% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vc.R
\name{write_vc}
\alias{write_vc}
\title{write_vc}
\usage{
write_vc(work_item, path_absolute, output_dir, export_as)
}
\arguments{
\item{work_item}{A data.frame or tibble with a single row, passed by drop_name()}

\item{path_absolute}{A logical parameter specifying, whether eventually a relative or absolute path should be returned.}

\item{output_dir}{A relative path (in regard to the working directory) where the visual citations should be 'dropped'. (type: character)}

\item{export_as}{Defines the file format of the returned visual citation (see drop_name() for more).}
}
\value{
The path to the written file as character.
}
\description{
Takes the data from drop_name and writes the actual outputs to the output directory.
}
\examples{
\dontrun{
# not intended for direct call. Please refer to the documentation
# of drop_name() for further assistance.
}

}
