% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_html.R
\name{drop_html}
\alias{drop_html}
\title{drop_html}
\usage{
drop_html(
  title = "",
  journal = "",
  authors = "",
  year = "",
  cite_key = "",
  include_qr = TRUE,
  style = "modern"
)
}
\arguments{
\item{title}{The publication's title as string.}

\item{journal}{The Journal's name as string.}

\item{authors}{The authors' names as string. If this is a list, it has to be collapsed to a single string (s. example).}

\item{year}{The publication's publication year as string.}

\item{cite_key}{A string specifying the citation key within the .bib file. If no key is specified, the first entry is used.}

\item{include_qr}{Boolean value, whether to include a QR code (containing the URL to the DOI) next to the visual citation.}

\item{style}{A string specifying the desired style for the visual citation. Possible values are: "modern", "classic", "clean", "none". If "none" is given, the returned html can use a custom css file provided by the user. The custom CSS file must specify the DIV classes "top-row", "title-row" and "author-row".}
}
\value{
A htmltools taglist containing the visual citation as HTML representation including style.
}
\description{
Accepts bibliographic information and retuns a htmltools taglist for printing/display.
}
\examples{
\dontrun{
vc_html <- drop_html(
  title = bibentry$title,
  cite_key = bibentry$key,
  journal = bibentry$journal,
  year = bibentry$year,
  authors = paste(bibentry$author, collapse = ", ")
)

htmltools::html_print(dropped)
}
}
