% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_name.R
\name{drop_name}
\alias{drop_name}
\title{drop_name}
\usage{
drop_name(
  file = "sample_data/sample.bib",
  cite_key = NULL,
  export_as = "html",
  max_authors = 3
)
}
\arguments{
\item{file}{A .bib file.}

\item{cite_key}{A string specifying the citation key within the .bib file. If no key is specified, the first entry is used.}

\item{export_as}{A string specifying the desired output format. For now only supports HTML by using "html".}

\item{max_authors}{Integer number of maximum authors to print. If the number of authors exceeds this, the list is cropped accordingly.}
}
\value{
A visual citation
}
\description{
Extracts metadata from a .bib file and exports the visual citation in the specified format.
}
\examples{
\dontrun{
dropped <- drop_name(
 file = "sample_data/sample.bib",
 cite_key = NULL,
 export_as = "html",
 max_authors = 3)

 htmltools::html_print(dropped)
 }
}
