% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_css_styles.R
\name{get_css_styles}
\alias{get_css_styles}
\title{get_css_style}
\usage{
get_css_styles(style)
}
\arguments{
\item{style}{A string specifying the desired style for the visual citation. Possible values are:
"modern", "classic", "clean", "none". If "none" is given, the returned html can use a custom css file provided by the user.
This custom CSS file must specify styles for <div> classes "top-row", "title-row" and "author-row".}
}
\value{
A list of inline css styles for each element of the visual citation: top row, title row and author row.
}
\description{
Provides inline CSS code for three distinct visual citation styles.
If "none" is given, the returned styles are empty strings.
}
