% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedropR-package.R
\docType{package}
\name{namedropR-package}
\alias{namedropR}
\alias{namedropR-package}
\title{namedropR: Create Visual Citations to Include in Presentations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides 'visual citations' containing the metadata of a scientific paper and a 'QR' code. A 'visual citation' is a banner containing title, authors, journal and year of a publication. This package can create such banners based on 'BibTeX' and 'BibLaTeX' references and includes a QR code pointing to the 'DOI'. The resulting HTML object or PNG image can be included in a presentation to point the audience to good resources for further reading. Styling is possible via predefined designs or via custom 'CSS'. This package is not intended as replacement for proper reference manager packages, but a tool to enrich scientific presentation slides.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nucleic-acid/namedropR}
  \item Report bugs at \url{https://github.com/nucleic-acid/namedropR/issues}
}

}
\author{
\strong{Maintainer}: Christian A. Gebhard \email{christian.gebhard@posteo.de} (\href{https://orcid.org/0000-0002-8792-565X}{ORCID})

}
\keyword{internal}
